refactor: ajoute un wrapper sur tous les composants liés au résultat
