+++ src/views/simulation/resultats/benefits-bafa-list.vue
@@ -1,17 +1,11 @@
 <script setup lang="ts">
 import BenefitsList from "@/components/benefits-list.vue"
-import { useRouter } from "vue-router"
-import BackButton from "@/components/buttons/back-button.vue"
 import { useResultsStore } from "@/stores/results.js"
 import { computed, onMounted } from "vue"
 import Simulation from "@/lib/simulation"
-import LoadingModal from "@/components/loading-modal.vue"
 import { BenefitGroup } from "@data/types/benefits"
 
-const router = useRouter()
 const resultsStore = useResultsStore()
-const fetching = computed(() => resultsStore.fetching)
-const updating = computed(() => resultsStore.updating)
 const benefits = computed(() => resultsStore.benefits)
 const groupBenefits = computed(() => {
   const bafaBafdGroup = resultsStore.benefitTree.find(
@@ -30,16 +24,6 @@ onMounted(async () => {
 <template>
   <div>
     <h1 class="fr-text--lead">Aides BAFA et BAFD</h1>
-    <LoadingModal v-if="fetching || updating">
-      <p>Récupération en cours…</p>
-    </LoadingModal>
-    <BackButton
-      class="fr-mb-4w"
-      data-testid="back-button"
-      size="small"
-      @click="router.push({ name: 'resultats' })"
-      >Retour aux résultats
-    </BackButton>
     <BenefitsList
       v-if="groupBenefits"
       :benefits-and-benefit-groups="groupBenefits"
