+++ src/views/simulation/resultats/resultats.vue
@@ -1,9 +1,4 @@
 <template>
-  <LoadingModal v-if="fetching || updating">
-    <p v-show="fetching"> Récupération de la situation en cours… </p>
-    <p v-show="updating"> Calcul en cours de vos droits… </p>
-  </LoadingModal>
-
   <ErrorsEmailAndSmsModal />
 
   <WarningMessage v-if="hasWarning">
@@ -91,28 +86,16 @@ import ErrorSaveBlock from "@/components/error-save-block.vue"
 import Feedback from "@/components/feedback.vue"
 import OfflineResults from "@/components/offline-results.vue"
 import TrouverInterlocuteur from "@/components/trouver-interlocuteur.vue"
-import LoadingModal from "@/components/loading-modal.vue"
-import StatisticsMixin from "@/mixins/statistics.js"
 import WarningMessage from "@/components/warning-message.vue"
 import Recapitulatif from "@/views/simulation/recapitulatif.vue"
-import { useStore } from "@/stores/index.js"
 import { useResultsStore } from "@/stores/results.js"
-import { daysSinceDate } from "@lib/utils.js"
-import { EventAction, EventCategory } from "@lib/enums/event.js"
 import ErrorsEmailAndSmsModal from "@/components/modals/errors-email-and-sms-modal.vue"
-import Simulation from "@/lib/simulation.js"
-import MockResults from "@/lib/mock-results"
-import { computed, onMounted, onBeforeUnmount, ref } from "vue"
-import { useRoute, useRouter } from "vue-router"
+import { computed } from "vue"
 
 const resultsStore = useResultsStore()
-const store = useStore()
-const router = useRouter()
-const route = useRoute()
 const benefits = computed(() => resultsStore.benefits)
 const benefitTree = computed(() => resultsStore.benefitTree)
 const hasWarning = computed(() => resultsStore.hasWarning)
-const fetching = computed(() => resultsStore.fetching)
 const updating = computed(() => resultsStore.updating)
 const error = computed(() => resultsStore.error)
 const hasErrorSave = computed(() => resultsStore.hasErrorSave)
@@ -120,164 +103,8 @@ const isSimulationUnavailable = computed(
   () => resultsStore.isSimulationUnavailable
 )
 const shouldDisplayResults = computed(() => resultsStore.shouldDisplayResults)
-const stopSubscription = ref<(() => void) | null>(null)
-
-onMounted(async () => {
-  initializeStore()
-  handleLegacySituationId()
-  if (MockResults.mockResultsNeeded()) {
-    MockResults.mock(route.params.benefitId)
-    return
-  } else if (route.query?.simulationId) {
-    await handleSimulationIdQuery()
-  } else if (!store.passSanityCheck) {
-    await restoreLatestSimulation()
-  } else if (store.calculs.dirty) {
-    await saveSimulation()
-  } else if (!store.hasResults) {
-    if (store.simulation.teleservice) {
-      await redirectToTeleservice()
-    } else {
-      store.computeResults()
-    }
-  }
-})
-
-onBeforeUnmount(() => {
-  stopSubscription.value?.()
-})
-
-const restoreLatestSimulation = async () => {
-  const lastestSimulationId = Simulation.getLatestId()
-  if (!lastestSimulationId) {
-    StatisticsMixin.methods.sendEventToMatomo(
-      EventCategory.General,
-      EventAction.Redirection,
-      route.path
-    )
-
-    return store.redirection((route) => router.push(route))
-  }
-
-  StatisticsMixin.methods.sendEventToMatomo(
-    EventCategory.General,
-    EventAction.CalculResultatsRestauration,
-    route.path
-  )
-
-  await store.fetch(lastestSimulationId)
-
-  if (store.simulationAnonymized) {
-    await store.retrieveResultsAlreadyComputed()
-  } else {
-    store.computeResults()
-  }
-}
 
 const isEmpty = (array) => {
   return !array || array.length === 0
-}
-const sendShowStatistics = () => {
-  StatisticsMixin.methods.sendBenefitsStatistics(
-    benefits.value,
-    EventAction.Show
-  )
-}
-const sendDisplayUnexpectedAmountLinkStatistics = () => {
-  const benefitsWithUnexpectedAmount = benefits.value.filter((benefit) => {
-    const unexpectedAmountLinkDisplayed =
-      (benefit.isBaseRessourcesYearMinusTwo &&
-        !store.ressourcesYearMinusTwoCaptured) ||
-      benefit.showUnexpectedAmount
-
-    return unexpectedAmountLinkDisplayed
-  })
-
-  StatisticsMixin.methods.sendBenefitsStatistics(
-    benefitsWithUnexpectedAmount,
-    EventAction.ShowUnexpectedAmountLink
-  )
-}
-
-const sendAccessToAnonymizedResults = () => {
-  StatisticsMixin.methods.sendEventToMatomo(
-    EventCategory.General,
-    EventAction.AccesSimulationAnonymisee,
-    daysSinceDate(new Date(store.simulation.dateDeValeur)).toString()
-  )
-}
-const initializeStore = () => {
-  store.updateCurrentAnswers(route.path)
-
-  stopSubscription.value = store.$onAction(({ after, name }) => {
-    after(() => {
-      switch (name) {
-        case "setResults": {
-          sendShowStatistics()
-          sendDisplayUnexpectedAmountLinkStatistics()
-          break
-        }
-        case "saveComputationFailure": {
-          StatisticsMixin.methods.sendEventToMatomo(
-            EventCategory.General,
-            EventAction.ErreurInitStore,
-            route.path
-          )
-          break
-        }
-      }
-    })
-  })
-}
-const handleLegacySituationId = () => {
-  // Used for old links containing situationId instead of simulationId
-  if (route.query?.situationId) {
-    store.setSimulationId(route.query.situationId.toString())
-  }
-}
-const handleSimulationIdQuery = async () => {
-  if (store.simulationId === route.query.simulationId && store.hasResults) {
-    return
-  }
-
-  if (route.query.simulationId) {
-    await store.fetch(route.query.simulationId.toString())
-  }
-
-  if (store.simulationAnonymized) {
-    sendAccessToAnonymizedResults()
-    await store.retrieveResultsAlreadyComputed()
-  } else {
-    store.computeResults()
-  }
-
-  router.replace({ ...router.currentRoute.value, simulationId: null } as any)
-}
-const saveSimulation = async () => {
-  try {
-    store.setSaveSituationError("")
-    await store.save()
-
-    if (!store.access.forbidden) {
-      store.computeResults()
-    }
-  } catch (error: any) {
-    store.setSaveSituationError(error.response?.data || error)
-    StatisticsMixin.methods.sendEventToMatomo(
-      EventCategory.General,
-      EventAction.ErreurSauvegardeSimulation,
-      route.path
-    )
-  }
-}
-const redirectToTeleservice = async () => {
-  if (store.simulationId) {
-    const representation = await store.fetchRepresentation(
-      store.simulation.teleservice,
-      store.simulationId
-    )
-
-    window.location.href = representation.destination.url
-  }
 }
 </script>
